+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 * double-click to select all blocks with the same predicates.
 * fix aura trait crash in some cases.
 * fix pnc pressure trait wont accept air.
-* improve forge energy bar
+* improve forge energy bar.
+* add pnc heat trait.
+* complete fuel recipe ui panel..